+++ src/lexer/action/action.ts
@@ -7,6 +7,7 @@ import type {
   AcceptedActionExecOutput,
 } from "./output";
 import { rejectedActionOutput, AcceptedActionOutput } from "./output";
+import { ActionWithKinds } from "./select";
 
 export type ActionExec<Data, ActionState, ErrorType> = (
   input: Readonly<ActionInput<ActionState>>,
@@ -351,5 +352,14 @@ export class Action<Data = never, ActionState = never, ErrorType = never> {
         Action.from<Data, ActionState, ErrorType>(a).or(b),
       ),
     );
+  }
+
+  /**
+   * Set kinds for this action. This is used if your action can yield multiple kinds.
+   * @example
+   * builder.branch(a => a.from(...).kinds(...).select(...))
+   */
+  kinds<Kinds extends string>(kinds: Kinds[]) {
+    return new ActionWithKinds(kinds, this);
   }
 }